Remove provider interface
