--- geth/notification/message/message.go
@@ -1,7 +1,0 @@
-package message
-
-// Message with data and payload
-type Message struct {
-	Body    interface{}
-	Payload *Payload
-}
